+++ src/common/resources/orders.js
@@ -1,0 +1,12 @@
+import {
+  baseManifest,
+  ORDERS_API,
+} from '@folio/stripes-acq-components';
+
+export const ordersResource = {
+  ...baseManifest,
+  path: ORDERS_API,
+  records: 'purchaseOrders',
+  accumulate: true,
+  fetch: false,
+};