+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## (IN PROGRESS)
 
+### Stories
+* [UINV-161](https://issues.folio.org/browse/UINV-161) Alert user when adding pol to invoice that has a different currency or vendor that the invoice
+
 ## [2.1.3]https://github.com/folio-org/ui-invoice/tree/v2.1.3) (2020-07-01)
 [Full Changelog](https://github.com/folio-org/ui-invoice/compare/v2.1.2...v2.1.3)